refactor: refactor dependencies
