+++ backend/app/api/dependencies/session.py
@@ -1,0 +1,8 @@
+from typing import Annotated
+
+from fastapi import Depends
+from sqlmodel import Session
+
+from app.core.database import get_session
+
+SessionDep = Annotated[Session, Depends(get_session)]
